+++ includes/cetech1/core/modules.h
@@ -1,6 +1,8 @@
 #pragma once
 #include "types.h"
 
+#define CT_DECL_MODULE(name, description) CT_MODULE_EXPORT uint8_t CT_CONCATENATE(ct_load_module_, name)
+
 typedef uint8_t(ct_module_fce_t)(
     const struct ct_apidb_api_t *api,
     const struct ct_allocator_t *allocator,