+++ src/cetech1/core/cetech1.zig
@@ -7,3 +7,6 @@ pub usingnamespace @import("kernel.zig");
 pub usingnamespace @import("log.zig");
 pub usingnamespace @import("bagraph.zig");
 pub usingnamespace @import("stringid.zig");
+pub usingnamespace @import("task.zig");
+
+pub usingnamespace @import("profiler.zig");