+++ src/cetech1/core/modules.zig
@@ -8,15 +8,15 @@ const LogAPI = @import("log.zig").LogAPI;
 pub const LoadModuleFn = fn (apidb: ?*const c.c.ct_apidb_api_t, _allocator: ?*const c.c.ct_allocator_t, load: u8, reload: u8) callconv(.C) u8;
 pub const LoadModuleZigFn = fn (apidb: *apidb.ApiDbAPI, allocator: std.mem.Allocator, _log: *LogAPI, load: bool, reload: bool) anyerror!bool;
 
-const LOG_SCOPE = "module";
-
 pub inline fn loadModuleZigHelper(comptime load_module_zig: LoadModuleZigFn, _apidb: ?*const c.c.ct_apidb_api_t, _allocator: ?*const c.c.ct_allocator_t, load: bool, reload: bool) u8 {
+    const MODULE_NAME = "module";
+
     var apidb_api = _apiFromCApi(_apidb.?);
     var allocator = _allocFromCApi(_allocator.?);
     var log = apidb_api.getZigApi(LogAPI).?;
 
     var r = load_module_zig(apidb_api, allocator, log, load, reload) catch |err| {
-        log.err(LOG_SCOPE, "Could not load module {}\n", .{err});
+        log.err(MODULE_NAME, "Could not load module {}\n", .{err});
         return 0;
     };