+++ src/cetech1/core/private/log.zig
@@ -2,6 +2,8 @@ const std = @import("std");
 
 const public = @import("../cetech1.zig");
 const apidb = @import("apidb.zig");
+const profiler = @import("profiler.zig");
+const task = @import("task.zig");
 
 pub var api = public.LogAPI{
     .log = log,
@@ -15,12 +17,29 @@ pub fn registerToApi() !void {
 }
 
 pub fn log(level: public.LogAPI.Level, scope: []const u8, msg: []const u8) void {
-    const LOG_FORMAT = "\t{s}: {s}";
+    if (profiler.profiler_enabled) {
+        const LOG_FORMAT_TRACY = "{s}: {s}";
+        var color: u32 = switch (level) {
+            .info => 0x00_ff_ff_ff,
+            .debug => 0x00_00_ff_00,
+            .warn => 0x00_ff_ef_00,
+            .err => 0x00_ff_00_00,
+        };
+        var buffer: [256:0]u8 = undefined;
+        _ = std.fmt.bufPrintZ(&buffer, LOG_FORMAT_TRACY, .{ scope, msg }) catch undefined;
+        profiler.api.msgWithColor(&buffer, color);
+    }
+
+    const thread_id = std.Thread.getCurrentId();
+    //const thread_name = task.getThreadName(thread_id);
+
+    const LOG_FORMAT = "\t[{d}] {s}:\t{s}";
+    const args = .{ thread_id, scope, msg };
     switch (level) {
-        .info => std.log.info(LOG_FORMAT, .{ scope, msg }),
-        .debug => std.log.debug(LOG_FORMAT, .{ scope, msg }),
-        .warn => std.log.warn(LOG_FORMAT, .{ scope, msg }),
-        .err => std.log.err(LOG_FORMAT, .{ scope, msg }),
+        .info => std.log.info(LOG_FORMAT, args),
+        .debug => std.log.debug(LOG_FORMAT, args),
+        .warn => std.log.warn(LOG_FORMAT, args),
+        .err => std.log.err(LOG_FORMAT, args),
     }
 }